Make practice overlay into individual class.
